requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
openai>=1.40.0
